pull config for other ansible hosts from their vars
Before this commit, we had to define all the hosts, including their IPs
and the public keys in the vars of the host we deployed to. If the other
hosts are using ansible for deployments too and are in the same
inventory, you can now just define them as peers using their
inventory_name instead, ansible will pull in the right vars for you. For
that to work you will need to specify their yggdrasil_listen_address
though.